refactor: invite dir 변경
