+++ src/components/invite/BackgroundView.tsx
