+++ src/components/invite/CardViewArea/CardPreview.tsx
