+++ src/components/invite/CardViewArea/GuideBox.tsx
