+++ src/components/invite/CardViewArea/InfoBox.tsx
