+++ src/components/invite/LoginModal/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const LoginModal = () => {
+  return <div></div>;
+};
+
+export default LoginModal;
